feat(containercontroller): add subscribe functionality to container
It was not at home in a usergroupcontroller.
