+++ src/index.js
@@ -1,11 +1,9 @@
 import Connection from './connection/connection';
 import ContainerController from './container/container-controller';
 import FillLevelController from './fill-level/fill-level-controller';
-import UsergroupController from './usergroup/usergroup-controller';
 
 export {
   Connection,
   ContainerController,
-  FillLevelController,
-  UsergroupController
+  FillLevelController
 };
